[RPA-1464] Getting required credential types from process
