+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/cloudFile/downloadCloudFile.action.ts
@@ -2,9 +2,11 @@ import { CloudFileAction, MicrosoftPlatform } from 'runbotics-common';
 
 import { IBpmnAction, Runner } from '#src-app/Actions/types';
 import { translate } from '#src-app/hooks/useTranslations';
+import { ActionCredentialType } from '#src-app/credentials/actionCredentialType.enum';
 
-export const downloadCloudFileAction: IBpmnAction = {
+export const downloadCloudFileAction = {
     id: CloudFileAction.DOWNLOAD_FILE,
+    credentialType: ActionCredentialType.MICROSOFT_GRAPH,
     label: translate('Process.Details.Modeler.Actions.CloudFile.Download.Label'),
     script: CloudFileAction.DOWNLOAD_FILE,
     runner: Runner.DESKTOP_SCRIPT,
@@ -118,4 +120,4 @@ export const downloadCloudFileAction: IBpmnAction = {
             },
         },
     },
-};
+} satisfies IBpmnAction;
