+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/cloudFile/getSharepointListItemsAction.ts
@@ -2,9 +2,11 @@ import { CloudFileAction, MicrosoftPlatform } from 'runbotics-common';
 
 import { IBpmnAction, Runner } from '#src-app/Actions/types';
 import { translate } from '#src-app/hooks/useTranslations';
+import { ActionCredentialType } from '#src-app/credentials/actionCredentialType.enum';
 
-export const getSharepointListItemsAction: IBpmnAction = {
+export const getSharepointListItemsAction = {
     id: CloudFileAction.GET_SHAREPOINT_LIST_ITEMS,
+    credentialType: ActionCredentialType.MICROSOFT_GRAPH,
     label: translate('Process.Details.Modeler.Actions.CloudFile.GetSharePointListItems.Label'),
     script: CloudFileAction.GET_SHAREPOINT_LIST_ITEMS,
     runner: Runner.DESKTOP_SCRIPT,
@@ -80,4 +82,4 @@ export const getSharepointListItemsAction: IBpmnAction = {
             },
         },
     },
-};
+} satisfies IBpmnAction;
