+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/cloudFile/uploadCloudFile.action.ts
@@ -2,9 +2,11 @@ import { CloudFileAction, MicrosoftPlatform } from 'runbotics-common';
 
 import { IBpmnAction, Runner } from '#src-app/Actions/types';
 import { translate } from '#src-app/hooks/useTranslations';
+import { ActionCredentialType } from '#src-app/credentials/actionCredentialType.enum';
 
-export const uploadCloudFileAction: IBpmnAction = {
+export const uploadCloudFileAction = {
     id: CloudFileAction.UPLOAD_FILE,
+    credentialType: ActionCredentialType.MICROSOFT_GRAPH,
     label: translate('Process.Details.Modeler.Actions.CloudFile.Upload.Label'),
     script: CloudFileAction.UPLOAD_FILE,
     runner: Runner.DESKTOP_SCRIPT,
@@ -94,4 +96,4 @@ export const uploadCloudFileAction: IBpmnAction = {
             },
         },
     },
-};
+} satisfies IBpmnAction;
