+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/googleSheets.actions.ts
@@ -1,6 +1,6 @@
 import { GoogleAction } from 'runbotics-common';
 
-import { ActionCredentialType } from '#src-app/Actions/action-credential-type';
+import { ActionCredentialType } from '#src-app/credentials/actionCredentialType.enum';
 import { translate } from '#src-app/hooks/useTranslations';
 
 import { IBpmnAction, Runner } from './types';
