+++ runbotics/runbotics-orchestrator-ui/src/src-app/Actions/types.ts
@@ -1,7 +1,7 @@
 import { UiSchema } from '@rjsf/core';
 import { JSONSchema7 } from 'json-schema';
 
-import { ActionCredentialType } from '#src-app/Actions/action-credential-type';
+import { ActionCredentialType } from '#src-app/credentials/actionCredentialType.enum';
 
 export enum Runner {
     BROWSER_BACKGROUND_SCRIPT = '${environment.services.backgroundScript()}',
