+++ runbotics/runbotics-orchestrator-ui/src/src-app/views/process/ProcessBuildView/ModelerPanels/RunSavePanel.tsx
@@ -14,7 +14,6 @@ import { ModelerErrorType } from '#src-app/store/slices/Process';
 import { StyledBotProcessRunner } from './ModelerPanels.styled';
 import TooltipError from './TooltipError';
 import FloatingGroup from '../FloatingGroup';
-import { getProcessRequiredCredentials } from '#src-app/views/process/ProcessBuildView/getProcessRequiredCredentials';
 
 interface RunSavePanelProps {
     process: IProcess;
@@ -31,8 +30,7 @@ const RunSavePanel: FC<RunSavePanelProps> = ({
     const { isSaveDisabled, errors, customValidationErrors } = useSelector(
         (state) => state.process.modeler
     );
-    getProcessRequiredCredentials();
-
+    
     const getTooltip = () => {
         const {
             formErrorElementsNames,
